domain.Position translated to double to handle mid road nodes

+++ algos/domain/Position.java
@@ -1,3 +1,3 @@
 package domain;
 
-public record Position(int x, int y) {}
+public record Position(double x, double y) {}
